컨트롤러 서비스 리포지토리 책임분리 리팩토링

+++ src/main/java/com/cloudingYo/barrierFree/place/repository/PlaceRepository.java
@@ -1,12 +1,11 @@
 package com.cloudingYo.barrierFree.place.repository;
 
-import com.cloudingYo.barrierFree.place.dto.PlaceCoordinateDTO;
+import com.cloudingYo.barrierFree.place.dto.resp.PlaceCoordinateDTO;
 import com.cloudingYo.barrierFree.place.entity.Place;
-import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
 public interface PlaceRepository {
-    Optional<Place> findByPlaceKey(int placeKey);
-    Optional<PlaceCoordinateDTO> findCoordinateByPlaceKey(int placeKey);
+    Optional<Place> findByPlaceKey(Long placeKey);
+    Optional<PlaceCoordinateDTO> findCoordinateByPlaceKey(Long placeKey);
 }
