+++ src/main/java/com/cloudingYo/barrierFree/place/repository/PlaceRepositoryImpl.java
@@ -1,8 +1,7 @@
 package com.cloudingYo.barrierFree.place.repository;
 
-import com.cloudingYo.barrierFree.place.dto.PlaceCoordinateDTO;
+import com.cloudingYo.barrierFree.place.dto.resp.PlaceCoordinateDTO;
 import com.cloudingYo.barrierFree.place.entity.Place;
-import com.cloudingYo.barrierFree.user.repository.UserJPARepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
@@ -15,12 +14,12 @@ public class PlaceRepositoryImpl implements PlaceRepository {
     private final PlaceJPARepository placeJPARepository;
 
     @Override
-    public Optional<Place> findByPlaceKey(int placeKey) {
+    public Optional<Place> findByPlaceKey(Long placeKey) {
         return placeJPARepository.findByPlaceKey(placeKey);
     }
 
     @Override
-    public Optional<PlaceCoordinateDTO> findCoordinateByPlaceKey(int placeKey) {
+    public Optional<PlaceCoordinateDTO> findCoordinateByPlaceKey(Long placeKey) {
         return placeJPARepository.findCoordinateByPlaceKey(placeKey);
     }
